pyproject.toml
[project]
name = "baby-monitor-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "python-telegram-bot>=22.0",
    "requests>=2.32.3",
]
